requirements.txt
astroid==2.3.3
certifi==2019.11.28
colorama==0.4.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
olefile==0.46
opencv-contrib-python==4.2.0.32
Pillow==7.0.0
pylint==2.4.4
typed-ast==1.4.0
wincertstore==0.2
wrapt==1.11.2
